flask_system/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0 